feat: add call to action block to home page
This commit adds a new call-to-action block to the home page. The `HomePageProps` type is updated to include a `callToAction` property, which consists of an image, title, content (HTML), and link (label and URL). The `HomeContentWidget` component is also updated to render the new call-to-action block using the `CallToActionBlock` component.