feat(api): add consult-booking-request route (#20)
* feat(blogs): update BlogsPage and BlogsContentWidget components

- Removed unused searchParams prop from BlogsPage component
- Removed PaginationButtonsBlock and SearchBoxBlock components
- Updated getAllBlogSummary function in read/index.ts to remove unnecessary parameters

* feat: add code changes for the book-a-consult page

This commit adds code changes to the book-a-consult page. It imports necessary modules and components, defines types, and implements functions to generate metadata and render the content widget and hero widget. Additionally, it includes a new query to retrieve data for the book-a-consult page.

* feat(api): add consult-booking-request route

This commit adds a new route for handling consult booking requests in the API. The route handles POST requests and includes validation and processing of the request data, as well as integration with reCAPTCHA for spam prevention.

feat(api/send): add consult-booking-request notification route

This commit adds a new route for sending email notifications when a consult booking request is made. The route uses the Resend library to send the email, including details of the booking request.

feat(sitemap): add book-a-consult URL to sitemap

This commit updates the sitemap to include a new URL for the "book-a-consult" page. The URL has a monthly change frequency and priority of 0.5.

feat(components/blocks): add consultation-form-block component

This commit adds a new component called "ConsultationFormBlock" that represents a form for submitting consultation requests. The component includes form validation using zod, integration with reCAPTCHA, and submission handling logic.

chore: update dependencies