chore: update dependencies (#30)
* feat(image-display-block): update image quality and sizes

- Update image quality to 100
- Adjust image sizes for different screen widths

* chore(deps): update @types/react to 18.2.57, @vercel/analytics to 1.2.1, embla-carousel-react to 8.0.0-rc23, and lucide-react to 0.334.0

* chore: update dependencies

* feat: Update font import and variable names

The code changes update the font import and variable names in the layout.tsx file. The Inter font is replaced with Montserrat, and the corresponding CSS variables are updated accordingly. Additionally, the tailwind.config.ts file is modified to use the new font variable for the sans font family.

* feat(solutions): add solutions page components and data retrieval

This commit adds the necessary components and functions to display the solutions page. It also includes a new function `getSolutionsPage` that retrieves the required data for the solutions page from the backend.

The following files were modified:
- app/(root)/(solutions)/solutions/page.tsx
- components/widgets/blogs-content-widget.tsx
- components/widgets/solutions-content-widget.tsx
- components/widgets/solutions-hero-widget.tsx
- lib/data/operations/queries/index.ts
- lib/data/read/index.ts

* feat(image-display-block): update image sizes

The `sizes` attribute in the `ImageDisplayBlock` component has been updated to improve responsiveness.

* chore: update dependencies