feat: add PAS header block to solution content widget (#32)
* feat(sitemap): add solutions to sitemap

This commit adds the ability to include solutions in the sitemap. It retrieves all solution summaries and generates URLs for each solution, including their slug, last modified date, change frequency, and priority.

* feat(widgets): import data read functions

This commit adds imports for the `getAllBlogSummary`, `getAllProjectSummary`, and `getAllSolutionSummary` functions in the respective widget files.

* feat: add solution page and solution hero widget

This commit adds a new solution page and a solution hero widget. The solution page imports the necessary components and defines the required props. It also includes an async function to generate metadata for the page. The solution hero widget receives props for title and heroActionBlock, which are used to display the hero section of the solution page.

The commit also includes a new query (qrySolutionBySlug) in the queries file, which retrieves a specific solution by its slug. Additionally, there is a new function (getSolutionBySlug) in the read file that fetches data for a specific solution using its slug.

These changes enhance the functionality of the application by allowing users to view individual solutions with their respective hero sections.

* feat: add PAS header block to solution content widget

This commit adds a new PASHeaderBlock component to the SolutionContentWidget in order to display the header section of the pasBlock. The PASHeaderBlock receives props for image, title, subtitle, and content from the pasBlock.