feat: add benefit card block component
This commit adds a new component called BenefitCardBlock, which is used to display a benefit card with a title and content. It also includes changes to import and use the BenefitCardBlock component in the SolutionContentWidget component.

The commit message follows the conventional commits specification, with a description of less than 50 characters and in lowercase.