feat(hero-display-block): update Separator background color (#42)
* feat: update dependencies and add border to hero display block

chore(package.json): update @types/node to version 20.11.20

chore(pnpm-lock.yaml): update @types/node to version 20.11.20

chore(pnpm-lock.yaml): update embla-carousel-react to version 8.0.0

chore(pnpm-lock.yaml): update lucide-react to version 0.336.0

* feat: Update privacy policy page and hero widget

This commit updates the privacy policy page and hero widget components. It adds a new property `secure_url` to the `image` object in both components, allowing for secure image URLs to be used. Additionally, it limits the number of FAQs displayed to 13 in the `qrySolutionBySlug` query.

* feat: update image object in ConsultPageProps and ConsultHeroWidget

The code changes include updating the image object in the ConsultPageProps and ConsultHeroWidget components to include a secure_url property. This ensures that the secure URL of the image is properly handled and displayed.

* feat(support): update support page and hero display block

- Updated the `support/page.tsx` file to include a new property `secure_url` in the `image` object.
- Modified the `hero-display-alternate-block.tsx` file to adjust the layout and styling of the hero display block.
- Updated the `support-hero-widget.tsx` file to use the new `secure_url` property for displaying images.

This commit improves the support page and enhances the visual presentation of hero display blocks.

* feat: Update image URLs in blog page components

This commit updates the image URLs in the `BlogPage` and `BlogsPage` components to use the `secure_url` property instead of the `public_id` property. It also updates the styling in the `HeaderDisplayBlock` and `HeroDisplayAlternateBlock` components to display white text on badges and subtitles.

* feat(testimonials): add secure_url to image object

This commit adds the `secure_url` property to the `image` object in both `testimonials/page.tsx` and `testimonials-hero-widget.tsx`.

* feat(projects): update image object with secure_url property

The code changes in this commit update the `image` object in the `ProjectPageProps` and `ProjectsPageProps` types, as well as in the `ProjectHeroWidgetProps` component. The `image` object now includes a new property called `secure_url`. This change allows for the use of secure URLs when displaying project images.

* feat: update image object in SolutionPageProps and SolutionsPageProps

This commit updates the image object in SolutionPageProps and SolutionsPageProps by adding the secure_url property.

* feat: refactor hero display blocks

- Refactored the HeroDisplayAlternateBlock component to use the HeroDisplayBlock component instead.
- Updated import statements in various files to reflect this change.

* feat(hero-display-block): update Separator background color

The Separator component in the HeroDisplayBlock now has a white background color.