feat: update HomePage component to include pasBlock (#48)
* feat(homepage): update home page layout and hero widget

This commit updates the layout of the home page by adjusting the spacing between sections. It also modifies the hero widget component to display a background image and improve the alignment of text and buttons. Additionally, it adds support for rendering HTML content within the hero widget.

Conventional Commits Specification:
- Type: feat
- Scope: homepage
- Description: update home page layout and hero widget

* feat: Update home page content and hero widget styling

The code changes include removing the "content" property from the HomePageProps type in page.tsx, updating the text size and width of the title in home-hero-widget.tsx, and removing the "content" property from qryHomePage query in index.ts.

* feat: update HomePage component to include pasBlock

This commit updates the HomePage component to include the pasBlock prop. It also adds the pasBlock prop to the HomeContentWidget and AboutContentWidget components.