feat: Add new components for home page (#52)
* feat: add testimonials to home page

This commit adds the necessary code changes to include testimonials on the home page. It includes updates to the `HomePageProps` and `HomeContentWidgetProps` types, as well as modifications to the `HomePage` and `HomeContentWidget` components. The changes also include updates to the GraphQL query for retrieving testimonials data.

* feat: Add new components for home page

This commit adds new components for the home page, including `BlogsCarouselBlock`, `HomeBenefitCardBlock`, and `HomeProcessCardBlock`. It also updates the `HomeContentWidget` to include these new components. Additionally, it adds a query to retrieve the latest blog summaries in the `qryLatestBlogSummary` function and a helper function `getLatestBlogSummary` to fetch this data.